Work around Waybar bug
Waybar makes method calls on the StatusNotifierItem's unique D-Bus name,
rather than its well-known name.  With the current code, this causes a
problem: as every StatusNotifierItem has the same path, it is not
possible to identify which StatusNotifierItem the message was intended
for.  Instead, use a separate object path for every StatusNotifierItem,
removing any ambiguity.  Sending a message to an incorrect well-known
name is explicitly detected and rejected.