Rip out dbusmenu support
The D-Bus menu code never actually worked.  Worse, the C library that it
would likely have been paired with, libdbusmenu, is horribly brittle
and easily fails assertions.  Therefore, it is not feasable to sanitize
data passed to it.

This will be replaced by having the guest render the menu into a
surface, which the host will position correctly via wlr-layer-shell.